+++ Yazmin/iff.h
@@ -27,12 +27,12 @@ unsigned int IFFID(char c1, char c2, char c3, char c4);
 
 unsigned int unpackLong(const void *data);
 
-void packLong(const void *data, unsigned int value);
+void packLong(const void *data, unsigned long value);
 
 int isID(const void *data, char c1, char c2, char c3, char c4);
 int isForm(const void *data, char c1, char c2, char c3, char c4);
 unsigned int chunkIDAndLength(const void *data, unsigned int *chunkID);
-unsigned int paddedLength(unsigned int len);
+unsigned long paddedLength(unsigned long len);
 
 void IFFCreateBuffer(IFFHandle *handle, unsigned long initialBufferSize);
